+++ src/main/java/org/evtstore/provider/TransactNeo4jProvider.java
@@ -121,7 +121,7 @@ public class TransactNeo4jProvider implements Provider {
         var ev = result.single().get("ev");
 
         var stored = event.clone();
-        stored.version = ev.get("version").asInt();
+        stored.version = ev.get("version").asDouble();
         stored.position = ev.get("position").asOffsetDateTime().format(DateTimeFormatter.ISO_DATE_TIME);
         trx = null;
         return stored;
@@ -140,7 +140,7 @@ public class TransactNeo4jProvider implements Provider {
       var ev = result.single().get("ev");
 
       var stored = event.clone();
-      stored.version = ev.get("version").asInt();
+      stored.version = ev.get("version").asDouble();
       stored.position = ev.get("position").asOffsetDateTime().format(DateTimeFormatter.ISO_DATE_TIME);
       return stored;
     } catch (ClientException ex) {
@@ -213,7 +213,7 @@ public class TransactNeo4jProvider implements Provider {
     ev.aggregateId = record.get("aggregateId").asString();
     ev.event = record.get("event").asString();
     ev.stream = record.get("stream").asString();
-    ev.version = record.get("version").asInt();
+    ev.version = record.get("version").asDouble();
     var timestamp = record.get("timestamp").asOffsetDateTime().format(DateTimeFormatter.ISO_DATE_TIME);
     var dt = OffsetDateTime.parse(timestamp);
     ev.timestamp = new Date(dt.toInstant().toEpochMilli());
